requirements.txt
gradio==3.0.24
editdistance
pyspellchecker
spacy==3.2.1
flask==2.1.3
gunicorn